バリデーションメッセージを日本語化
勘定科目削除用のバリデーション追加

+++ app/Http/Controllers/AccountController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Account;
+use App\Journal;
 
 class AccountController extends Controller
 {
-    private $validation = array(
+
+    private $storeValidation = array(
         'account' => 'required|unique:accounts|max:20',
         'type' => 'required'
     );
@@ -20,7 +22,7 @@ class AccountController extends Controller
 
     public function store(Request $request)
     {
-        $this->validate($request, $this->validation);
+        $this->validate($request, $this->storeValidation);
         $account = new Account();
         $account->account = $request->account;
         $account->type = $request->type;
@@ -43,8 +45,16 @@ class AccountController extends Controller
         //
     }
 
-    function destroy(Request $request, $id){
+    function destroy(Request $request, $id)
+    {
         $account = Account::find($id);
+        $journals = new journal();
+        $debit = $journals->where('debit', $account->id)->get();
+        $credit = $journals->where('credit', $account->id)->get();
+
+        if ($debit->count() or $credit->count()) {
+            return response()->json('この勘定科目は仕訳で使用されているので削除できません。');
+        }
         $account->delete();
         return response()->json();
     }
